requirements.txt
numpy
jinja2
matplotlib
imgkit
beautifulsoup4
ffmpeg-python
scipy
